This commit include the following new features
- New default nand value for 512MB size
- Correct rqsm6duallite machine name
- Fixed enumeration mmc device for rqs7 modules
- Bug fix linux script q7_tftp_kernel.sh

+++ recipes-core/engicam-emmc-script/engicam-emmc-script/q7_tftp_kernel.sh
@@ -1,26 +1,30 @@
-serverip=192.168.2.58
-filename=uImage
-dtb=uImage.dtb
-
-
 if env | grep -q ^serverip=; then
- 	echo "Download " $filename
+	filename=uImage
+	dtb1=imx6dl-icore-rqs.dtb
+	dtb2=imx6q-icore-rqs.dtb
+	echo "Download " $filename "-" $dtb1 "/" $dtb2
 	tftp -g -r $filename -l $filename $serverip
-	tftp -g -r $dtb -l $dtb  $serverip
 	if [ -f $filename ]; then
-        mount -t vfat /dev/mmcblk0p1 /mnt/
-	mv $filename /mnt/$filename
-        mv $dtb /mnt/$dtb
-	sync
-	umount /mnt
-else
-	echo "Unable to download " $filename
-fi
-
+		tftp -g -r $dtb1 -l $dtb1  $serverip
+		if [ -f $dtb1 ]; then
+		    tftp -g -r $dtb2 -l $dtb2  $serverip
+		    if [ -f $dtb2 ]; then
+			mount -t vfat /dev/mmcblk0p1 /mnt/
+			mv $filename /mnt/$filename
+			mv $dtb1 /mnt/$dtb1
+			mv $dtb2 /mnt/$dtb2
+			sync
+			umount /mnt
+		    else
+		      echo "Unable to download " $dtb2
+		    fi
+		else
+		  echo "Unable to download " $dtb1
+		fi
+	else
+		echo "Unable to download " $filename
+	fi
 else
 	echo "Serverip not setted:"
 	echo "    export serverip=192.168.XXX.XXX"
 fi
-
-
-
